+++ src/main/java/com/iaraapi/repository/AccessTypeRepository.java
@@ -3,5 +3,5 @@ package com.iaraapi.repository;
 import com.iaraapi.model.AccessType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccessTypeRepository extends JpaRepository<AccessType, Long> {
+public interface AccessTypeRepository extends JpaRepository<AccessType, Integer> {
 }
