+++ src/main/java/com/iaraapi/repository/FactoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FactoryRepository extends JpaRepository<Factory,Long> {
+public interface FactoryRepository extends JpaRepository<Factory,Integer> {
 }
